+++ pjecz_casiopea_flask/config/settings.py
@@ -26,6 +26,7 @@ class Settings(BaseSettings):
     TASK_QUEUE_NAME: str = os.getenv("TASK_QUEUE_NAME", "pjecz_casiopea")
     TZ: str = os.getenv("TZ", "America/Mexico_City")
     TURNOS_API_KEY: str = os.getenv("TURNOS_API_KEY", "")
+    TURNOS_API_KEY_URL: str = os.getenv("TURNOS_API_KEY_URL", "")
 
     # Incrementar el tamaño de lo que se sube en los formularios
     MAX_CONTENT_LENGTH: int | None = None